feat: cleanup file structure
