+++ packages/1ui/src/components/SidebarLayout/components/SidebarLayout.tsx
@@ -1,0 +1,10 @@
+import { ResizablePanelGroup } from '@components/Resizable'
+
+export const SidebarLayout = ({ ...props }) => {
+  return (
+    <ResizablePanelGroup
+      direction="horizontal"
+      {...props}
+    ></ResizablePanelGroup>
+  )
+}
