+++ packages/1ui/src/components/SidebarLayout/components/SidebarLayoutContent.tsx
@@ -1,0 +1,12 @@
+import { ResizablePanel } from '@components/Resizable'
+
+export const SidebarLayoutContent = ({ ...props }) => {
+  return (
+    <ResizablePanel defaultSize={70}>
+      <div
+        className="flex h-full items-center justify-center p-3"
+        {...props}
+      ></div>
+    </ResizablePanel>
+  )
+}
