+++ packages/1ui/src/components/SidebarLayout/components/SidebarLayoutNavFooter.tsx
@@ -1,0 +1,8 @@
+export const SidebarLayoutNavFooter = ({ ...props }) => {
+  return (
+    <div
+      className="flex h-full w-full flex-col justify-end gap-2 p-2"
+      {...props}
+    ></div>
+  )
+}
