+++ packages/1ui/src/components/SidebarLayout/components/SidebarLayoutNavFooterItem.tsx
@@ -1,0 +1,18 @@
+import { SidebarNavItemProps, SidebarNavItem } from './SidebarNavItem'
+
+export const SidebarLayoutNavFooterItem = ({
+  iconName,
+  label,
+  onClick,
+  ...props
+}: SidebarNavItemProps) => {
+  return (
+    <SidebarNavItem
+      {...props}
+      variant="text"
+      label={label}
+      iconName={iconName}
+      onClick={onClick}
+    />
+  )
+}
