+++ packages/1ui/src/components/SidebarLayout/components/SidebarLayoutNavHeader.tsx
@@ -1,0 +1,8 @@
+export const SidebarLayoutNavHeader = ({ ...props }) => {
+  return (
+    <div
+      className="border-primary/30 flex w-full items-center gap-2 border-0 border-b-[1px] border-solid p-4"
+      {...props}
+    ></div>
+  )
+}
