+++ packages/1ui/src/components/SidebarLayout/components/SidebarLayoutNavHeaderContent.tsx
@@ -1,0 +1,19 @@
+import { useSidebarLayoutContext } from './SidebarLayoutProvider'
+
+export interface SidebarLayoutNavHeaderContentProps {
+  imgLogo: React.ReactElement
+  textLogo: React.ReactElement
+}
+
+export const SidebarLayoutNavHeaderContent = ({
+  imgLogo,
+  textLogo,
+}: SidebarLayoutNavHeaderContentProps) => {
+  const { isCollapsed } = useSidebarLayoutContext()
+  return (
+    <>
+      {imgLogo}
+      {!isCollapsed && textLogo}
+    </>
+  )
+}
