+++ packages/1ui/src/components/SidebarLayout/components/SidebarLayoutNavItem.tsx
@@ -1,0 +1,18 @@
+import { SidebarNavItemProps, SidebarNavItem } from './SidebarNavItem'
+
+export const SidebarLayoutNavItem = ({
+  iconName,
+  label,
+  onClick,
+  ...props
+}: SidebarNavItemProps) => {
+  return (
+    <SidebarNavItem
+      {...props}
+      variant="ghost"
+      label={label}
+      iconName={iconName}
+      onClick={onClick}
+    />
+  )
+}
