+++ packages/1ui/src/components/SidebarLayout/components/SidebarLayoutNavItems.tsx
@@ -1,0 +1,3 @@
+export const SidebarLayoutNavItems = ({ ...props }) => {
+  return <div className="flex w-full flex-col gap-2 p-2" {...props}></div>
+}
