+++ packages/1ui/src/components/SidebarLayout/components/index.ts
@@ -1,0 +1,11 @@
+export * from './SidebarLayoutProvider'
+export * from './SidebarLayout'
+export * from './SidebarLayoutNav'
+export * from './SidebarLayoutNavHeader'
+export * from './SidebarLayoutNavHeaderContent'
+export * from './SidebarLayoutContent'
+export * from './SidebarLayoutNavFooter'
+export * from './SidebarLayoutNavFooterItem'
+export * from './SidebarLayoutNavItem'
+export * from './SidebarLayoutNavItems'
+export * from './SidebarLayoutNavAvatar'
