+++ packages/1ui/src/components/SidebarLayout/constants.ts
@@ -1,0 +1,1 @@
+export const SIDEBAR_LOCAL_STORAGE_VARIABLE = 'isSideBarCollapsed'
