+++ packages/1ui/src/components/SidebarLayout/index.ts
@@ -1,1 +1,1 @@
-export * from './SidebarLayout'
+export * from './components'
